reformat file structures + set nav in each component
