+++ src/About/index.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import Nav from '../Navbar';
+
+const About = () => {
+
+  return ( 
+    <>
+    <Nav/>
+    <div>Put about description here</div>
+    </>
+  )
+}
+
+export default About;
