+++ src/Navbar/NavStyles.js
@@ -1,11 +1,11 @@
 import styled from 'styled-components';
 
 export const Navbar = styled.div`
-  padding: 8px;
+  ${'' /* padding: 8px; */}
   display: flex;
   height: 50px;
   align-items: center;
   justify-content: left;
-  position: absolute;
+  ${'' /* position: absolute; */}
   z-index:1;
 `;
