--- src/components/About.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const About = () => {
-  return <div>Put about description here</div>
-}
-
-export default About;
